requirements.txt
discord.py==2.0.1
ffmpeg==1.4
python-dotenv==0.20.0
yt-dlp==2022.03.08.1
requests==2.27.1
youtube-dl==2021.12.17
PyNaCl==1.5.0
apscheduler==3.9.1
discord-py-interactions=4.3.2